pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "domstar-northstar"
version = "0.1.0"
description = "DOM-grounded training and evaluation pipeline for Northstar CUA Fast."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.1.0",
    "bitsandbytes>=0.45.0",
    "datasets>=3.1.0",
    "kernel",
    "numpy>=1.26.0",
    "peft>=0.18.0",
    "pillow>=10.4.0",
    "playwright>=1.49.0",
    "scikit-learn>=1.5.0",
    "transformers>=4.57.0",
    "trl>=0.23.0",
    "tzafon",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
